renamed: PopulationChartページ| contextフォルダ名変更
先頭にフォルダが移動するようにアンダーバーを頭文字に
