+++ src/pages/PopulationChartPage/_contexts/PrefectureColorPicker/type.ts
